+++ src/sink/mod.rs
@@ -1,0 +1,5 @@
+mod sink_option;
+mod sink_result;
+
+pub use sink_option::SinkOption;
+pub use sink_result::SinkResult;