+++ CHANGELOG.md
@@ -4,12 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## Unreleased
+
+## [0.6.1] - 10/31/2025
 
 ### Added
 
-- Added `Sink` trait
+- Added `SinkOption` trait
 - Added `ThenNone` trait - for transforming a `bool` into an `None` when true
+- Added `SinkResult` trait
 
 ### Removed